Fix the year calculating of the entry date of a transaction
In the section :61: there are two dates value date and entry date
value date has a year but entry date not.

We now try to caculate the year of the entry date by expecting that all of
these dates have to be quiet near to another.
So if value date is in december and entry date is in january, the entry
date has to be in the next year.